Make tests pass Delete stuff we don't need
