+++ server/test/non-mocha/test.ahk
@@ -1,0 +1,3 @@
+TestFunc() {
+	MsgBox "HI"
+}
